moved the container aware HTTP kernel to the HttpKernel component
